+++ rbx_binary/src/serializer.rs
